Store database and table details in one place
The previous commits to this repository had major problems: If the database details were different, the administrator had to change the details in each file. That is time consuming and prone to errors.
Therefore, I present a major update in which all the database and table details are stored in one place. This is also useful if the administrators change the name of each table in the database (usually for security reasons).
This commit also resolves some existing bugs in the source code.